Changed int to double Added images for weather

+++ Examensarbete/Models/ApiModels/DarkskyModels/Alert.cs
@@ -10,8 +10,8 @@ namespace Examensarbete.Models.ApiModels.DarkskyModels
         public string title { get; set; }
         public List<string> regions { get; set; }
         public string severity { get; set; }
-        public int time { get; set; }
-        public int expires { get; set; }
+        public double time { get; set; }
+        public double expires { get; set; }
         public string description { get; set; }
         public string uri { get; set; }
     }
