+++ Examensarbete/Models/ApiModels/DarkskyModels/Currently.cs
@@ -5,8 +5,8 @@
         public int time { get; set; }
         public string summary { get; set; }
         public string icon { get; set; }
-        public int precipIntensity { get; set; }
-        public int precipProbability { get; set; }
+        public double precipIntensity { get; set; }
+        public double precipProbability { get; set; }
         public double temperature { get; set; }
         public double apparentTemperature { get; set; }
         public double dewPoint { get; set; }
@@ -14,9 +14,9 @@
         public double pressure { get; set; }
         public double windSpeed { get; set; }
         public double windGust { get; set; }
-        public int windBearing { get; set; }
+        public double windBearing { get; set; }
         public double cloudCover { get; set; }
-        public int uvIndex { get; set; }
+        public double uvIndex { get; set; }
         public double visibility { get; set; }
         public double ozone { get; set; }
     }
